OpenWB 2: fix negative phase readings

+++ charger/openwb-2.0.go
@@ -182,7 +182,7 @@ func (wb *OpenWB20) getPhaseValues(reg uint16) (float64, float64, float64, error
 
 	var res [3]float64
 	for i := range res {
-		res[i] = float64(binary.BigEndian.Uint16(b[2*i:])) / 100
+		res[i] = float64(int16(binary.BigEndian.Uint16(b[2*i:]))) / 100
 	}
 
 	return res[0], res[1], res[2], nil
